Create Efficient LinkedList in Python
Current Implementation of Linked List in Python requires for creation of a separate class, called Node, so that they can be connected using a main Linked List class. In the provided implementation, the Linked List is created without defining a separate class for a node. Using the proposed implementation, Linked Lists are easier to understand and can be simply visualized using the print function.